Pull the `DeterministicRandom` library from its repo.
Using no `:from` clause on a dependency is deprecated,
and all standard library packages are moving to their own repos.

The only reason we didn't do this in the initial commit is
because we had to release this `Random` library first, as
`DeterministicRandom` has a dependency on `Random`.
(Note that the `Random` library only depends on `DeterministicRandom`
for its `spec` manifest, but not the real `Random` library manifest).